fix performance issue while displaying active track
When track was updated (by tracker usually)
it was added to the map as a new overlay object.
But the old overlay object was left there.
So after longer time when tracker was active,
map holds multiple versions of one track.
It had huge performance and memory impact.

This commit fixing that. It is re-using
current segment ids for overlay objects.